requirements.txt
absl-py==0.7.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.7.1
astroid==2.2.2
attrs==18.1.0
beautifulsoup4==4.7.1
Brlapi==0.6.6
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cognitive-face==1.4.2
command-not-found==0.3
ConfigArgParse==0.14.0
coursera-dl==0.11.4
cryptography==2.6.1
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
defusedxml==0.5.0
distro-info==0.18
dj-database-url==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-crispy-forms==1.7.2
djangorestframework==3.9.2
emoji==0.5.2
entrypoints==0.3
Flask==1.0.2
gast==0.2.2
grpcio==1.19.0
gunicorn==19.9.0
h5py==2.9.0
httplib2==0.9.2
idna==2.8
isort==4.3.10
itsdangerous==1.1.0
jeepney==0.4
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==19.0.1
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mock==2.0.0
netifaces==0.10.4
newsapi-python==0.2.3
numpy==1.16.2
oauth==1.0.1
oauthlib==3.0.1
olefile==0.45.1
opencv-python==4.0.0.21
pandas==0.24.2
pbr==5.1.3
pexpect==4.2.1
Pillow==5.4.1
protobuf==3.7.1
psycopg2==2.8.2
pyasn1==0.4.5
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.7.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.3.1
pyRFC3339==1.0
PySocks==1.6.8
python-apt==1.6.3+ubuntu1
python-dateutil==2.8.0
python-debian==0.1.32
python-dotenv==0.10.1
python3-openid==3.1.0
pytz==2018.9
pyxdg==0.25
PyYAML==5.1
reportlab==3.4.0
requests==2.21.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
scikit-learn==0.20.3
scipy==1.2.1
SecretStorage==3.1.1
simplejson==3.13.2
six==1.12.0
sklearn==0.0
soupsieve==1.9
system-service==0.3
systemd-python==234
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
twilio==6.26.0
typed-ast==1.2.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
uuid==1.30
virtualenv==16.4.0
wadllib==1.3.2
Werkzeug==0.15.2
whitenoise==4.1.2
wrapt==1.11.1
xkit==0.0.0
zope.interface==4.3.2
